Add a GCP bastion module, move the AWS module to a sub-directory, use a Docker image for `terraform-docs` (#22)
* Add a GCP bastion module, move the AWS module to a sub-directory, use a Docker image for `terraform-docs`

The new GCP Terraform module manages a Google Cloud bastion, using Google DNS for the bastion host record, Stackdriver for logs, and Google Cloud Storage for SSH hostkeys and optional `additional external users`.

The AWS bastion module is moved to its own `aws` sub-directory - the next release of the AWS module will require users to update their `source` argument to include the `//aws` path and a module-specific release tag such as `aws-vx.y.z`.

A Docker image is used to incorparate `terraform-docs` output into per-module ReadMe files, until a version of terraform-docs is released that natively supports Terraform 0.12.
* Terraform-docs pull request: https://github.com/segmentio/terraform-docs/pull/113
* Docker image and Terraform 0.12 workaround: https://github.com/cytopia/docker-terraform-docs/pull/8

* Update GCP firewall rule to be optional, use the bastion name for the FW rule tag

The GCP firewall rule is optionally created, if the `ssh_cidr_blocks`
input is set to an empty list.

The GCP instance template and firewall rule use the bastion name as the
tag that associates the firewall rule with the bastion compute instance.

* Update GCP module providers to be inline with other related 3rd party modules

* GCP module additional external users GCS object is optional, other fixes

The additional external users script is only written to GCS if the
`additional_external_users` input is defined.

References to `user data` were renamed to `setup script` in template resources.

The Google provider version is further relaxed to 2.X.

* Update the GCP `ssh_public_key_file` input to behave like the AWS module

The `ssh_public_key_file` input was initially given an unfortunate name, and actually reflects the content
(not the file name) of an SSH public key. For now, the GCP module will
match the AWS one.